+++ client/Packages/com.beamable.server/Editor/UI/Components/PublishPopup/PublishPopup.cs
@@ -5,6 +5,7 @@ using Beamable.Editor.UI.Components;
 using Beamable.Editor.UI.Model;
 using Beamable.Server.Editor;
 using Beamable.Server.Editor.UI.Components;
+using UnityEditor;
 using UnityEngine;
 #if UNITY_2018
 using UnityEngine.Experimental.UIElements;
@@ -40,15 +41,17 @@ namespace Beamable.Editor.Microservice.UI.Components
             }
         }
 
-        public ManifestModel Model { get; set; }
         public Action OnCloseRequested;
         public Action<ManifestModel> OnSubmit;
 
+        public ManifestModel Model { get; set; }
+        
         private TextField _generalComments;
         private Button _cancelButton;
         private PrimaryButtonVisualElement _continueButton;
         private ScrollView _scrollContainer;
-        private List<PublishManifestEntryVisualElement> _publishManifestElements;
+        private Dictionary<string, PublishManifestEntryVisualElement> _publishManifestElements;
+        private bool _isPublishDisabled;
 
         public PublishPopup() : base(nameof(PublishPopup))
         {
@@ -68,14 +71,41 @@ namespace Beamable.Editor.Microservice.UI.Components
                 return;
             
             _scrollContainer = Root.Q<ScrollView>("manifestsContainer");
-            _publishManifestElements = new List<PublishManifestEntryVisualElement>(Model.Services.Count);
+            _publishManifestElements = new Dictionary<string, PublishManifestEntryVisualElement>(Model.Services.Count);
             
-            foreach (var kvp in Model.Services)
+            List<IEntryModel> entryModels = new List<IEntryModel>(Model.Services.Values);
+            entryModels.AddRange(Model.Storages.Values);
+
+            bool isOddRow = true;
+            _isPublishDisabled = false;
+            foreach (var model in entryModels)
             {
-                var newElement = new PublishManifestEntryVisualElement {Model = kvp.Value};
+                if (!MicroserviceConfiguration.Instance.EnableStoragePreview && model is StorageEntryModel)
+                {
+                    continue;
+                }
+                
+                if (MicroserviceConfiguration.Instance.EnableStoragePreview)
+                {
+                    DisablePublishFeature("Warning! In order to publish services you must disable Storage Preview first.");
+                }
+
+                if (model is ManifestEntryModel)
+                {
+                    var serviceModel = MicroservicesDataModel.Instance.GetModel<MicroserviceModel>(model.Name);
+                    if (serviceModel != null && !(serviceModel is RemoteMicroserviceModel) && serviceModel.Descriptor.IsPublishFeatureDisabled())
+                    {
+                        DisablePublishFeature("Warning! Publish feature is disabled due to Microservices dependent on Storage Objects.");
+                    }
+                }
+                
+                bool wasPublished = EditorPrefs.GetBool(GetPublishedKey(model.Name), false);
+                var newElement = new PublishManifestEntryVisualElement(model, wasPublished, isOddRow);
                 newElement.Refresh();
-                _publishManifestElements.Add(newElement);
+                _publishManifestElements.Add(model.Name, newElement);
                 _scrollContainer.Add(newElement);
+
+                isOddRow = !isOddRow;
             }
 
             _generalComments = Root.Q<TextField>("largeCommentsArea");
@@ -86,6 +116,7 @@ namespace Beamable.Editor.Microservice.UI.Components
 
             _continueButton = Root.Q<PrimaryButtonVisualElement>("continueBtn");
             _continueButton.Button.clickable.clicked += () => OnSubmit?.Invoke(Model);
+            _continueButton.SetEnabled(!_isPublishDisabled);
         }
 
         public void PrepareForPublish()
@@ -94,14 +125,14 @@ namespace Beamable.Editor.Microservice.UI.Components
             _generalComments.RemoveFromHierarchy();
             _continueButton.RemoveFromHierarchy();
             _cancelButton.RemoveFromHierarchy();
-            for (int i = 0; i < _publishManifestElements.Count; i++)
-                _publishManifestElements[i].RemoveFromHierarchy();
+            foreach (var kvp in _publishManifestElements)
+                kvp.Value.RemoveFromHierarchy();
             _publishManifestElements.Clear();
 
             
             foreach (var kvp in Model.Services)
             {
-                var microserviceModel = MicroservicesDataModel.Instance.GetModel<MicroserviceModel>(kvp.Value.ServiceName);
+                var microserviceModel = MicroservicesDataModel.Instance.GetModel<MicroserviceModel>(kvp.Value.Name);
 
                 if (microserviceModel == null)
                 {
@@ -114,6 +145,19 @@ namespace Beamable.Editor.Microservice.UI.Components
                 _scrollContainer.Add(newElement);
                 new DeployMSLogParser(newElement, microserviceModel, true);
             }
+        }
+
+        public void ServiceDeployed(IDescriptor descriptor)
+        {
+            EditorPrefs.SetBool(GetPublishedKey(descriptor.Name), true);
+        }
+
+        private string GetPublishedKey(string serviceName) => string.Format(Microservices.SERVICE_PUBLISHED_KEY, serviceName);
+
+        private void DisablePublishFeature(string message)
+        {
+            Root.Q<Label>("warningMessage").text = message;
+            _isPublishDisabled = true;
         }
     }
 }